+++ streamlit_app/app.py
@@ -1,7 +1,42 @@
 import streamlit as st
+import json
+import requests
+from src.app_utils import *
+from src.raid import Raid
+import pandas as pd
 
-from app_utils import *
-from swgoh_guild_webapp.entities import Guild
+
+@st.cache_data(show_spinner=False, ttl=24*60*60)
+def get_raid_data(fetcher_url: str, guild_id: str, raid_id: str, interval_days: int) -> pd.DataFrame:
+    response = requests.get(f'{fetcher_url}/raid/{guild_id}?raid_id={raid_id}&interval_days={interval_days}')
+    df_dict = json.loads(response.content)
+    return Raid(pd.DataFrame(df_dict))
+
+
+def app_startup() -> None:
+    st.set_page_config(layout="wide")
+    start_vars = {
+        # 'allycode': '000000000',
+        'guild_id': 'dYXen85NS3SCrdllQ4lAEg',
+        'swgoh_comlinK_fetcher_url': 'https://swgoh-comlink-fetcher-4hzooxs5za-uc.a.run.app',
+        'raid': 'kraytdragon',
+        'raid_interval': 30, # days
+        'out_path': '/mnt/d/data/swgoh/guilds',
+        'op_unit_slack': 3, # this var is to denote the "danger zone" for unit availability vs planetary op need
+        'op_fig_char_max': 12, # max num of chars allowed in ops figures
+        'rote_data_path': '../data/rote_data', # relative path to rote data for main app
+        'rote_data_path_pages': '../../data/rote_data', # relative path to rote data for app pages
+    }
+    for var in start_vars:
+        if var not in st.session_state:
+            st.session_state[var] = start_vars[var]
+    with st.spinner('Fetching guild data. This may take a bit...'):
+        # st.session_state.guild_manager = init_guild(st.session_state.guild_id, st.session_state.out_path)
+        st.session_state.raid_data = get_raid_data(st.session_state['swgoh_comlinK_fetcher_url'],
+                                                   st.session_state['guild_id'],
+                                                   st.session_state['raid'],
+                                                   st.session_state['raid_interval'])
+    return
 
 
 def draw_guild_roster_view() -> None:
